pyproject.toml
[tool.poetry]
name = "bililivemedalcrawler"
version = "0.1.0"
description = ""
authors = ["cyanife"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
uvloop = "^0.16.0"
python-rapidjson = "^1.4"
aiohttp = "^3.7.4"
asks = "^2.4.12"
trio = "^0.19.0"

[tool.poetry.dev-dependencies]
pylint = "^2.10.2"
black = "^21.7b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
